requirements.txt
kafka-python==2.0.2
pandas==1.5.3
numpy==1.26.4
pyspark==3.5.1

